restructure evaluation
