+++ ads/aqua/evaluation/request_params.py
@@ -1,0 +1,87 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+# Copyright (c) 2024 Oracle and/or its affiliates.
+# Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl/
+
+"""
+aqua.evaluation.request_params
+~~~~~~~~~~~~~~
+
+This module contains requests parameters for aqua evaluation apis.
+"""
+
+from dataclasses import dataclass
+from typing import List, Optional
+
+from ads.common.serializer import DataClassSerializable
+
+
+@dataclass(repr=False)
+class CreateAquaEvaluationDetails(DataClassSerializable):
+    """Dataclass to create aqua model evaluation.
+
+    Fields
+    ------
+    evaluation_source_id: str
+        The evaluation source id. Must be either model or model deployment ocid.
+    evaluation_name: str
+        The name for evaluation.
+    dataset_path: str
+        The dataset path for the evaluation. Could be either a local path from notebook session
+        or an object storage path.
+    report_path: str
+        The report path for the evaluation. Must be an object storage path.
+    model_parameters: dict
+        The parameters for the evaluation.
+    shape_name: str
+        The shape name for the evaluation job infrastructure.
+    memory_in_gbs: float
+        The memory in gbs for the shape selected.
+    ocpus: float
+        The ocpu count for the shape selected.
+    block_storage_size: int
+        The storage for the evaluation job infrastructure.
+    compartment_id: (str, optional). Defaults to `None`.
+        The compartment id for the evaluation.
+    project_id: (str, optional). Defaults to `None`.
+        The project id for the evaluation.
+    evaluation_description: (str, optional). Defaults to `None`.
+        The description for evaluation
+    experiment_id: (str, optional). Defaults to `None`.
+        The evaluation model version set id. If provided,
+        evaluation model will be associated with it.
+    experiment_name: (str, optional). Defaults to `None`.
+        The evaluation model version set name. If provided,
+        the model version set with the same name will be used if exists,
+        otherwise a new model version set will be created with the name.
+    experiment_description: (str, optional). Defaults to `None`.
+        The description for the evaluation model version set.
+    log_group_id: (str, optional). Defaults to `None`.
+        The log group id for the evaluation job infrastructure.
+    log_id: (str, optional). Defaults to `None`.
+        The log id for the evaluation job infrastructure.
+    metrics: (list, optional). Defaults to `None`.
+        The metrics for the evaluation.
+    force_overwrite: (bool, optional). Defaults to `False`.
+        Whether to force overwrite the existing file in object storage.
+    """
+
+    evaluation_source_id: str
+    evaluation_name: str
+    dataset_path: str
+    report_path: str
+    model_parameters: dict
+    shape_name: str
+    block_storage_size: int
+    compartment_id: Optional[str] = None
+    project_id: Optional[str] = None
+    evaluation_description: Optional[str] = None
+    experiment_id: Optional[str] = None
+    experiment_name: Optional[str] = None
+    experiment_description: Optional[str] = None
+    memory_in_gbs: Optional[float] = None
+    ocpus: Optional[float] = None
+    log_group_id: Optional[str] = None
+    log_id: Optional[str] = None
+    metrics: Optional[List] = None
+    force_overwrite: Optional[bool] = False
